Fix fast refresh
