--- apps/samenvvv/components/HashtagAnnouncement/test.js
@@ -1,2 +1,0 @@
-export * from './types'
-export * from './HashtagAnnouncement'
