+++ libs/ui/src/components/Pagination/dots.ts
@@ -1,0 +1,1 @@
+export const DOTS = '...'
