+++ libs/ui/src/hooks/timeLocale.ts
@@ -1,0 +1,3 @@
+import { enIN as en, nl, tr } from 'date-fns/locale'
+
+export const timeLocale = { en, nl, tr }
