+++ libs/ui/src/hooks/useLocaleTimeFormat.ts
@@ -1,9 +1,9 @@
 import * as dateFns from 'date-fns'
-import { enIN as en, nl, tr } from 'date-fns/locale'
 import { useRouter } from 'next/router'
 
 import { StrapiLocale } from '@wsvvrijheid/types'
-export const timeLocale = { en, nl, tr }
+
+import { timeLocale } from './timeLocale'
 
 export const useLocaleTimeFormat = (time: string, format: string) => {
   const { locale } = useRouter()
